resolved comments except onBoarding swiper package
