+++ src/interfaces/drawer.ts
@@ -1,0 +1,13 @@
+import { DrawerNavigationProp } from '@react-navigation/drawer';
+
+type RootDrawerParamList = {
+  News: undefined;
+  Channel: undefined;
+  OverView: undefined;
+};
+
+type ScreenNavigationProp = DrawerNavigationProp<RootDrawerParamList, 'News'>;
+
+export interface DrawerProps {
+  navigation: ScreenNavigationProp;
+}
