+++ src/screen/drawerScreen/channels/index.tsx
@@ -1,24 +1,9 @@
 import React from 'react';
-import { DrawerNavigationProp } from '@react-navigation/drawer';
-
 import { View, Text, Button } from 'native-base';
 
-type RootDrawerParamList = {
-  News: undefined;
-  Channel: undefined;
-  OverView: undefined;
-};
+import { DrawerProps } from '../../../interfaces';
 
-type ProfileScreenNavigationProp = DrawerNavigationProp<
-  RootDrawerParamList,
-  'News'
->;
-
-type Props = {
-  navigation: ProfileScreenNavigationProp;
-};
-
-export default function Channel(props: Props) {
+export default function Channel(props: DrawerProps) {
   return (
     <View flex={1} alignItems={'center'} justifyContent={'center'}>
       <Text>Channel Screen q</Text>
